[ADD] new module: mail_template_conditional_report
This module is used to add condition (domain) on dynamic reports during generation of mail.template.
When the mail is generated (based on the mail template), all dynamic attachments are generated (by standard Odoo) then the domain is applied (based on the record) to know if the attachment should be dropped or not.